refactor(csr): modularize Send page, unify layout, and remove /app/api route
- Refactored the `/transfer/send` page to use `next/dynamic` for full client-side rendering (CSR).
- Modularized the localStorage dev page into a dedicated `LocalStorage` component with dynamic import.
- Improved loading performance by code-splitting large components and deferring rendering to the client.
- Unified layout logic across `/transfer` and `/dev` routes for cleaner structure and maintainability.
- Removed unused `/app/api` route and associated dummy file download handler.
- Cleaned up related utilities and removed redundant code.
- Minor UI and logging text refinements for consistency.

The app now uses **100% Client-Side Rendering (CSR)** powered by Next.js dynamic imports, optimized for performance and deployable on static hosting platforms.
