+++ relayr-ui/next.config.ts
@@ -1,8 +1,13 @@
 /** @type {import('next').NextConfig} */
 import type { NextConfig } from "next";
 
+const withBundleAnalyzer = require("@next/bundle-analyzer")({
+  enabled: process.env.ANALYZE === "true",
+});
+
 const nextConfig: NextConfig = {
   /* config options here */
+  output: "export",
 };
 
-module.exports = nextConfig;
+module.exports = withBundleAnalyzer(nextConfig);
