+++ relayr-ui/src/app/transfer/receive/page.tsx
@@ -1,25 +1,17 @@
-// React
-import { Suspense } from "react";
+"use client";
+
+// Next.js
+import dynamic from "next/dynamic";
 
 // Custom Components
 import ExperienceLoading from "@/components/loading/ExperienceLoading";
 
-// Internal Components
-import ReceivePageContent from "./ReceivePageContent";
+// Lazy-loaded Internal Components
+const Receive = dynamic(() => import("./Receive"), {
+  loading: () => <ExperienceLoading />,
+  ssr: false,
+});
 
-/**
- * File Transfer Page for the Seceiver
- * This component serves as the main entry point for the file receiver page,
- * wrapping the ReceivePageContent in a Suspense boundary
- * to handle loading states.
- *
- * @returns JSX.Element The receiver page content wrapped in a Suspense boundary.
- *
- */
-export default function ReceiverPage() {
-  return (
-    <Suspense fallback={<ExperienceLoading />}>
-      <ReceivePageContent />
-    </Suspense>
-  );
+export default function ReceivePage() {
+  return <Receive />;
 }
