+++ relayr-ui/src/app/transfer/send/page.tsx
@@ -1,45 +1,17 @@
 "use client";
 
-// File Transfer Components
-import SenderFlow from "@/components/file-transfer/sender/SenderFlow";
-import TransferCardLayout from "@/components/file-transfer/TransferCardLayout";
+// Next.js
+import dynamic from "next/dynamic";
 
-// Hooks
-import { useInitId } from "@/hooks/useInitId";
+// Custom Components
+import ExperienceLoading from "@/components/loading/ExperienceLoading";
 
-// Context Providers
-import { useSenderWebSocket } from "@/providers/SenderWebSocketProvider";
+// Lazy-loaded Internal Components
+const Send = dynamic(() => import("./Send"), {
+  loading: () => <ExperienceLoading />,
+  ssr: false,
+});
 
-// State Management (Store)
-import { useFileSenderStore } from "@/stores/useFileSenderStore";
-
-/**
- * File Transfer Page for the Sender
- * This page initializes the file transfer process for the sender,
- * displaying the sender's flow and managing the connection state.
- *
- * @returns JSX.Element The sender's file transfer page component.
- */
-export default function SenderPage() {
-  const { readyState } = useSenderWebSocket();
-
-  const file = useFileSenderStore((state) => state.file);
-  const { isTransferCompleted } = useFileSenderStore(
-    (state) => state.transferStatus,
-  );
-  const errorMessage = useFileSenderStore((state) => state.errorMessage);
-  const connectionId = useInitId("sender");
-
-  const showConnectionStatus = !!file || isTransferCompleted;
-
-  return (
-    <TransferCardLayout
-      readyState={readyState}
-      errorMessage={errorMessage}
-      connectionId={connectionId}
-      showConnectionStatus={showConnectionStatus}
-    >
-      <SenderFlow />
-    </TransferCardLayout>
-  );
+export default function SendPage() {
+  return <Send />;
 }
